Move still active button to post header
