+++ src/components/PostCard/PostBody/index.js
@@ -1,3 +1,2 @@
 import PostBody from './PostBody'
-import connector from './PostBody.connector'
-export default connector(PostBody)
+export default PostBody
