+++ src/components/PostCard/PostCompletion/PostCompletion.js
@@ -9,22 +9,21 @@ const promptOptions = {
   project: 'Is this still active?'
 }
 
-const otherMessageOptions = [
-  { label: 'Available', value: false },
-  { label: 'Expired', value: 'expired' },
-  { label: 'Unavailable', value: true }
-]
 const messages = {
   request: [
     { label: 'This is still needed', value: false },
-    { label: 'No longer needed', value: 'expired' },
-    { label: 'This request was completed', value: true }
+    { label: 'No longer needed', value: true }
   ],
-  offer: otherMessageOptions,
-  resource: otherMessageOptions,
+  offer: [
+    { label: 'Available', value: false },
+    { label: 'Unavailable', value: true }
+  ],
+  resource: [
+    { label: 'Available', value: false },
+    { label: 'Unavailable', value: true }
+  ],
   project: [
     { label: 'Active', value: false },
-    { label: 'Inactive', value: 'expired' },
     { label: 'Completed', value: true }
   ]
 }
@@ -38,9 +37,6 @@ export default function PostCompletion ({ type, startTime, endTime, isFulfilled,
       break
     case true:
       label = messages[type].find(choice => choice.value === true).label
-      break
-    case 'expired':
-      label = messages[type].find(choice => choice.value === 'expired').label
       break
     default:
       label = 'This is still needed'
