+++ src/components/PostCard/PostHeader/PostHeader.connector.js
@@ -4,6 +4,7 @@ import { removePostFromUrl, editPostUrl } from 'util/navigation'
 import getMe from 'store/selectors/getMe'
 import {
   deletePost,
+  unfulfillPost,
   fulfillPost,
   removePost,
   pinPost,
@@ -39,6 +40,9 @@ export function mapDispatchToProps (dispatch, props) {
     fulfillPost: postId => props.fulfillPost
       ? props.fulfillPost(postId)
       : dispatch(fulfillPost(postId)),
+    unfulfillPost: postId => props.unfulfillPost
+      ? props.unfulfillPost(postId)
+      : dispatch(unfulfillPost(postId)),
     removePost: postId => props.removePost
       ? props.removePost(postId)
       : dispatch(removePost(postId, groupSlug)),
@@ -51,7 +55,7 @@ export function mapDispatchToProps (dispatch, props) {
 export function mergeProps (stateProps, dispatchProps, ownProps) {
   const { currentUser, group } = stateProps
   const { id, creator } = ownProps
-  const { deletePost, editPost, fulfillPost, removePost, pinPost } = dispatchProps
+  const { deletePost, editPost, fulfillPost, unfulfillPost, removePost, pinPost } = dispatchProps
   const isCreator = currentUser && creator && currentUser.id === creator.id
   const canEdit = isCreator
   const canModerate = currentUser && currentUser.canModerate(group)
@@ -63,6 +67,7 @@ export function mergeProps (stateProps, dispatchProps, ownProps) {
     deletePost: isCreator ? () => deletePost(id, group ? group.id : null) : undefined,
     editPost: canEdit ? () => editPost(id) : undefined,
     fulfillPost: isCreator ? () => fulfillPost(id) : undefined,
+    unfulfillPost: isCreator ? () => unfulfillPost(id) : undefined,
     canFlag: !isCreator,
     pinPost: canModerate && group ? () => pinPost(id, group.id) : undefined,
     removePost: !isCreator && canModerate ? () => removePost(id) : undefined,
